ARM: sched: clear SD_SHARE_POWERDOMAIN
The ARM platforms take advantage of packing tasks on few cores if they can be
powergated independantly. We use DT to define at which level a core can be
independantly powergated to the others and the SD_SHARE_POWERDOMAIN will be set
accordingly at MC and CPU sched_domain level.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>