ARM: sched: clear SD_SHARE_POWERDOMAIN
The ARM platforms take advantage of packing tasks on few cores if they can be
powergated independantly. We use DT and the cpu topology method that is define
by these patches <add url> to define at which level a core can be
independantly powergated to the others and the SD_SHARE_POWERDOMAIN will be set
accordingly at MC and CPU sched_domain level.

The power-gate properties should be added with the value 1 in cpu and cluster
node when then can power gate independantly from the other.

As an example of a quad cores system which can power gate each core
independantly, we should have a DT similar to the example below

	cpus {
		#address-cells = <1>;
		#size-cells = <0>;

	       cpu-map {
	               cluster0 {
				power-gate = <1>;
				core0 {
	                               cpu = <&cpu0>;
					power-gate = <1>;
				};
				core1 {
				        cpu = <&cpu1>;
					power-gate = <1>;
				};
				core2 {
					cpu = <&cpu2>;
					power-gate = <1>;
				};
				core3 {
					cpu = <&cpu3>;
					power-gate = <1>;
				};
			};
		};

	...

	};

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>