configs: linaro-base: set CONFIG_AEABI=y
We already have CONFIG_THUMB2_KERNEL=y which also selects CONFIG_AEABI
however if someone wanted to disable Thumb2 for debugging or other
reasons then they may try simply removing the Thumb2 config entry. This
however would then default to the old ABI and it seems there are some
obscure bugs in this little tested kernel variant. So, to save wasting
people's time let's make sure CONFIG_AEABI is explicitly selected.

As we are adding this to linaro-base.conf we can remove it from
android.conf (which probably had it as a leftover from when we
had Thumb2 disabled).

Signed-off-by: Jon Medhurst <tixy@linaro.org>