Provide MSI controller registration mechanism
MSI support is currently provided through architecture call backs. This makes it
difficult to write architecture agnostic MSI controllers which exist outside arch/.
An example of this is in providing support for MSI controllers which will run on arm
and arm64.

At present each architecture implements the MSI callbacks and provide their own
abstraction and registration mechanism for providing MSI support for their platforms.
In some cases it's the host bridge drivers that implement the drivers/pci/msi.c MSI
callbacks.

This patch provides a common mechanism for all platforms to register MSI support. It
is hoped that further development of this mechansim will allow for multiple MSI
controllers serving different PCI domains and/or host-bridges.

This patch preserves the existing arch_ callback mechanism such that MSI support will
not break where this registration mechanism is not used - though it is hoped that
architectures will transition to this new mechanism and the old callbacks can be
deprecated.

The existing mechanism resulted in linker errors when arch_setup_msi_irq /
arch_teardown_msi_irq were not defined - this no longer occurs and a WARN is issued
during runtime.