drivers/gic: Use a private mapping for CPU target interfaces.
The GIC interface numbering does not necessarily follow the logical
CPU numbering, especially for complex topologies such as multi-cluster
systems.

Fortunately we can easily probe the GIC to create a mapping as the
Interrupt Processor Targets Registers for the first 32 interrupts are
read-only, and each field returns a value that always corresponds to
the processor reading the register.

Initially all mappings target all CPUs in case an IPI is required to
boot secondary CPUs.  It is refined as those CPUs discover what their
actual mapping is.

This patch is adapted to work with the new GIC driver that is going to
live in drivers/irqchip.

Signed-off-by: Nicolas Pitre <nico@linaro.org>