sched: Convert to state machine callbacks
The scheduler sports quite a bunch of hotplug notifiers. One reason
for multiple notifiers is the fact, that the startup and teardown
process are asymmetric. Now the scheduler wants to be called early on
startup and late on teardown. That requires to install two different
notifiers for the same issue.

With the state machine implementation we can register a callback pair
for startup and teardown at the appropriate spot.

This patch converts the notifiers which are setup with special
priorities and combines CPU_PRI_SCHED and CPU_PRI_CPUSET notifiers to
a single callback. They run back to back anyway and we can make sure
in the callbacks that the ordering inside the scheduler is
correct. These notifiers are installed in sched_init_smp() as we can't
run them during the bring up of the non boot CPUs because the SMP
scheduler is setup after that. CPUHP_SCHED_ONLINE is installed at
migration_init() and this happens before the others are installed
at sched_init_smp(). This callback should not run (cpu_stopper's kthread
is not yet setup) but we need this callback ready installed.

It would be nice if we just could compile them in, but that needs a
larger surgery to the scheduler code and is beyond the scope of this patch.

Reviewed-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>