ARM: Thermal: Support ARM PVT sensor
This patch adds support for the PVT sensor found on ARM TC2 testchip
and integrates with the thermal framework. The driver uses the device
tree description to discover the sensor.

The driver can provide trip points as part of a thermal zone to the
thermal framework and also supports generic cpufreq based cooling
devices described in the device tree. When the cooling devices are
provided, the framework can use them for thermal management.

The device tree node for the PVT sensor and it's properties are
described in the bindings documentation included with this patch.

Signed-off-by: Punit Agrawal <punit.agrawal@arm.com>
Change-Id: I6919a195aad8c30e0f4197674191174eb0c65022