pci: Add support for Juno GIC-MSI extension block
This driver provides support for the 'm' part of the future GICv2m block. As I
understand Juno won't have a GICv2m but instead a GICv2 and separate 'm'.

I've posted this on linux-eng for general feedback and also for alignment with
anyone working on similar/future GIC incarnations.

Bindings:

        msi@030000000 {
                compatible = "arm,gic-msi";

                reg = <0x0 0x2c1c0000 0 0x1000
                       0x0 0x2c1d0000 0 0x1000
                       0x0 0x2c1e0000 0 0x1000
                       0x0 0x2c1f0000 0 0x1000>;
        };

There is no way at present to determine how many frames are supported so
there is a reg for each.