Cpuidle: Add a new scheduled governor
This patch introduces a new cpuidle governor. The governor selects a state
based on precomputed idle period length and records the chosen state for use
by the scheduler.

Also included is additional code into kernel scheduler that does the
precomputation.

This patch is work in progress and is made availabe for
informational purposes only. Its purpose is to provoke comments and
discussion. The patch is not intended to be committed to any tree as-is
and certainly no-one should attempt to upstream it anywhere. If this
patch (in it's current form) is part of any such upstreaming proposal
by mistake, please NAK this. :)

Signed-off-by: Tuukka Tikkanen <tuukka.tikkanen@linaro.org>