[sched/core/cbs] Add policy definition for class::put_prev_task
This is an initial implementation for the release of a SE which has been
descheduled from a CPU into the CBS runqueue. The idea is pretty simple:
- run_cbs_entity_stop: keep track of the actual execution time
- put_prev_cbs_entity: update RQ data get back the SE in the round list

NOTE: the class::post_schedule handler is available only with SMP support.
Thus we prefer to keep SE timekeeping into methods which are always
available.