[core/sched/cbs] Controller: add entry points
The CBS controller will operate in three main points:
1. SE entering CPU
2. SE exiting CPU
3. New round start
These three entry points are hooked into the CBS internal API and they will
host the code in charge to tune and monitor controller specific metrics.
Specifically:
1. tune_cbs_burst: setup the next burst time to be assigned to a SE, once
the SE has been selected to enter the CPU;
2. monitor_cbs_burst: collects metrics related to the SE which are used to
update its status for the next run;
3. tune_cbs_round: setup the next scheduling round based on metrics collected
during the execution of the SE of the previous round.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>