[core/sched/cbs] Controller: add requoting support
Each time a SE enter/exit the RQ, the CPU time quota each SE is eligible to
get must be updated. Such a quota depends on the SE weight and the overall
RQ weight.

NOTE: the new quota should be operative starting from the next round

A couple of flags has been added to the RQ to track the need to requote while
the actual requoting is a two step operation:
1. at next round start, requoting base (i.e. 1/weight(RQ)) is cached
and the requoting flags properly udpated;
2. at each next burst of this round the actual SE quota is updated based
on previously cached requoting base value.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>