[core/sched/cbs] Controller: add burst monitoring support
Once a SE is removed from the CPU, its effective burst time is cached and
the round time updated. These values are required for the tuning of the
next round. This code is part of the monitor_cbs_burst CBS controller API
which is called for each SE right after they leave the CPU.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>