[core/sched/cbs] Controller: add reinitialization support
This adds the controller reinitialization code which is triggered each time
the RQ status change (e.g. SE enter/exit or prio change). A controller flag
is used to track the reinitialization condition, which is reset at each new
round if during the paste one there was not a reinitialization event.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>