[core/sched/cbs] Bugfix SE put/pick on sched_class changes
Once the RQ of a running task is changed we should consider the case of
such a task entering the CBS scheduler.
In that case, the task has not been scheduler for execution by the CBS
thus it is safer to request a new rescheduling to properly trigger the
CBS policy.

These changes are triggered so far by
- core::rt_mutex_setptio
- core::__sched_setscheduler
- core::sched_move_task
with all of them using this pattern:
1. put_prev_task
2. // updated scheduler and/or RQ
3. set_curr_task
4. enqueue_task

This patch provides this support:
A. trigger a rescheduling once a _running_ SE enter a CBS RQ
   => this allows the CBS policy to get control on the execution of this
      task and other RQ siblings
B. put just previously CBS scheduled entities
   indeed, if a running task is moved into CBS there is not need to stop
   and put the CBS entity

Thus, each time we enter the run_cbs_entity_stop and put_prev_cbs_entity
we are granted to being putting a previously CBS scheduled entity.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>