[core/sched/cbs] Bugfix CBS round error computation
The round error should be referred to the round time SP, which depends on
the number of active tasks in the RQ.

NOTE: the SP is updated _BEFORE_ computing the round error, which could result
into a different value depending on the tasks number variation during this
round. Perhaps we should introduce an SP_OLD value.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>