[core/sched/cbs] Bigfix Round Quota (RQ) computation
Because of the way load weights values are defined, and their range, the
pre-computation of the RQ load inverse using a RNQ_SCALE of just 12 bits
is not safe and is going to produce a null inverse.

Fixed point computations requires a re-design to better consider CFS defined
weights, which CBS want to preserve, their ranges and possible scaling and
overflow issues.

For the time being this patch disregard possible optimizations and keeps
just the division during SE round quota computation.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>