[core/sched/cbs] Bugfix RoundTime SP update
The SP for a round time should be updated _after_ the round error of the
previous run has been computed but _before_ the round event is traced via
ftrace, since such event should report the SP for the _next_ round.

This patch move the SP update into the round tuning function, right before
the next round event is going to be traced via FTrace.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>