[core/sched/cbs] Reset initialization after event tracing
The Burst FTrace event is generated after the task has completed its burst,
i.e. in the monitor_cbs_burst function.
This event reports the status of the re-initialization flag during last
burst assignment, thus this flag must be cleared after the burst completed.

This movement introduces a bit of run-time overhead since the flag is going
to be cleared even if the task was not re-initialized.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>