[core/sched/cbs] Optimization for queued HRTimer ticks
Task ticks calls generated by a scheduler HRTimer are notified to the
scheduling policy via the "queued" flag. In that case, we can avoid to
brother with preemption validation and just return with a resched request.

At the same time, when an HRTImer is pending, we can ignore periodic ticks
and return without rescheduling the current task, which is still consuming
its allotted burst.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>