[core/sched/cbs] Bugfix Reinit on Yield
Once a task is yielding the CPU, a reinitialization should be performed
during next burst tuning time. The actual simple solution tried to exploit
the re-init flags, which unfortunately is cleaned at burst monitoring time,
which is a call following the yield syscall.

This patch avoid the overloaded usage of the reinit flag by adding a new
status flag dedicated to track the yielding condition. This flag:
1. track the yielding condition from the yield syscall to the following
burst tuning call
2. allows the on monitor call to to properly report a clean reinit flag
for the current yielded burst
3. force the reinit flag on the next burst tuning call

This is an overall view of the proposed yield flag tracking:

TuneTime         YieldTime   MonitorTime      TuneTime
+------------------------+                    +--------+
|                        |                    |
+                        +-------+------------+
                                                   ^
                         +-------------------->    |
                                 +          +      |
                         Yielding|          |      |
                                 |          |      |
                                 |          +------+
                                 |
                                 |          Reinit
                                 v

                              Trace Yflag

NOTE: a queued task should not get Yielding set, since dequeuing could
also happen just because of a task migration.
The proper time to clear the yielding flag is just on the start of a new
burst, which correspond to suppose a task is not going to yield its CPU.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>