sched: Introduce energy data structures
The struct sched_group_energy represents the per scheduler group related
data which is needed for the energy aware scheduler.

It contains an atomic reference counter for scheduler internal
bookkeeping of data allocation and freeing, the number of elements and a
pointer to the idle state array which comprises (power consumption and
wakeup energy for the run->sleep->run cycle for this idle state) tuples
as well as the number of elements and a pointer to the capacity state
array which comprises (compute capacity, power consumption at this
compute capacity) tuples.

Allocation and freeing of struct sched_group_energy uses the existing
infrastructure of the scheduler which is currently used for the other sd
hierarchy data structures (e.g. struct sched_domain) as well.  That's why
struct sd_data is provisioned with a per cpu struct sched_group_energy
double pointer.

The struct sched_group obtains a pointer to a struct sched_group_energy.

The function ptr sched_domain_energy_f is introduced into struct
sched_domain_topology_level which will allow the arch to pass a particular
struct sched_group_energy from the topology shim layer into the scheduler
core.

The function pointer sched_domain_energy_f has an 'int cpu' parameter
since the folding of two adjacent sd levels via sd degenerate doesn't work
for all sd levels.  I.e. that it is not possible for example to use this
feature to provide per-cpu energy in sd level DIE on ARM's TC2 platform.

It was discussed that the folding of sd levels approach is preferable
over the cpu parameter approach, simply because the user (the arch
specifying the sd topology table) can introduce less errors.  But since
it is not working, the 'int cpu' parameter is the only way out.  It's
possible to use the folding of sd levels approach for
sched_domain_flags_f and the cpu parameter approach for the
sched_domain_energy_f at the same time though.  With the use of the
'int cpu' parameter, an extra check function has to be provided to make
sure that all cpus spanned by a sched group are provisioned with the same
energy data.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>