ARM: topology: Define sd energy and provide it to scheduler
!!! This patch is only there to be able to test provisioning of sd energy
related data from an arch topology shim layer to the scheduler.  Since
there is no code today which deals with extracting sd energy related data
from the dtb file and process it in the topology shim layer, the struct
sd_energy and the related struct capacity_state arrays are hard-coded here
!!!

This patch defines the struct sd_energy and the related struct
capacity_state array for the cluster (relates to sg's in DIE sd level)
and for the core (relates to sg's in MC sd level) for a cortex-a7 as well
as for a cortex-a15.  It further provides related implementations of the
sched_domain_energy_f functions.

To be able to propagate this information from the topology shim layer to
the scheduler, the elements of the arm_topology[] table have been
provisioned with the appropriate sched_domain_energy_f functions.

The cpu mask function cpu_corepower_mask() has been changed in such a way
that the sd degenerate code makes sure that in the (MC/GMC) sd level the
cortex-a7 sd energy data is attached to sg's related to the cortex-a7
CPU's whereas the cortex-a15 sd energy data is attached to to sg's related
to the to the cortex-a15 CPU's.

The cpu mask function cpu_cpupower_mask() has been added to create a
scenario with the appropriate counterpart cpu mask function cpu_cpu_mask()
in which the sd degenerate code makes sure that in the (DIE/GDIE) sd
level the cortex-a7 sd energy data is attached to sg's related to the
cortex-a7 CPU's whereas the cortex-a15 sd energy data is attached to to
sg's related to the to the cortex-a15 CPU's.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>